go: move all days into their own packages
I am becoming tired of sometimes remembering to prefix functions with
the day number, sometimes not. Having all days in their own packages
might be a little cumbersome due to import paths, but it should make
actually writing the solution code a little more convenient.

For now, all the code has simply been moved; it remains to refactor
and rename stuff.
