+++ go/internal/year2015/day01/day01.go
@@ -1,4 +1,4 @@
-package year2015
+package day01
 
 import (
 	"bufio"
