+++ go/internal/year2015/day04/day04.go
@@ -1,4 +1,4 @@
-package year2015
+package day04
 
 import (
 	"bufio"
