+++ go/internal/year2015/day04/day04_test.go
@@ -1,4 +1,4 @@
-package year2015
+package day04
 
 import (
 	"testing"
