+++ go/internal/year2015/day05/day05.go
@@ -1,4 +1,4 @@
-package year2015
+package day05
 
 import (
 	"fmt"
