+++ go/internal/year2015/day05/day05_test.go
@@ -1,4 +1,4 @@
-package year2015
+package day05
 
 import (
 	"testing"
