+++ go/internal/year2015/day06/day06_test.go
@@ -1,4 +1,4 @@
-package year2015
+package day06
 
 import (
 	"testing"
